+++ custom_components/recalbox/const.py
@@ -5,6 +5,6 @@ JSMODULES = [
     {
         "name": "Recalbox Card",
         "filename": "recalbox-card.js",
-        "version": "1.0.3",
+        "version": "1.1.0",
     },
 ]